backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
numpy==1.26.2
scikit-learn==1.3.2
hnswlib==0.8.0
networkx==3.2.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
redis==5.0.1
pydantic-settings==2.1.0
python-json-logger==2.0.7
google-generativeai==0.8.3